Using Leaderboard SetData to find current level
Pulse suggested this would be the ideal way to do it since SS also does it. Also, I did not use affinity patching as I wasnt sure if affinity could replace my other patches as they are all on internal types. So this is also going to be harmony for consistency sake.